new setup file
