+++ .github/workflows/python-test.yml
@@ -50,4 +50,4 @@ jobs:
       run: |
         pip install -e .
         pip install pytest
-        python -m pytest --import-mode=append MultiTrain_2/test/
+        python -m pytest --import-mode=append MultiTrain/test/
