+++ MultiTrain/__init__.py
@@ -1,3 +1,1 @@
 from MultiTrain.classification.classification_models import MultiClassifier
-from MultiTrain.regression.regression_models import MultiRegressor
-from MultiTrain.__version__ import __version__
