--- MultiTrain/__version__.py
@@ -1,3 +1,0 @@
-"""version information"""
-
-__version__ = "0.13.11"
