--- MultiTrain/tests/test_functions.py
@@ -1,5 +1,0 @@
-from MultiTrain.classification.classification_models import MultiClassifier
-
-
-def test_write_to_excel():
-    pass
