--- MultiTrain_2/__init__.py
@@ -1,1 +1,0 @@
-from MultiTrain.classification.classification_models import MultiClassifier
